start tear out motor
