+++ firmware/controllers/actuators/dc_motors.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cstddef>
+#include "global.h"
+
+class DcMotor;
+class Logger;
+
+DcMotor* initDcMotor(size_t index DECLARE_ENGINE_PARAMETER_SUFFIX);
+void showDcMotorInfo(Logging* logger);
